Initial migration to new app folder.
