+++ src/notes2notion/__main__.py
@@ -1,11 +1,5 @@
-#!/usr/bin/env python3
+"""Main interface to notes2notion."""
 
-# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED,
-# INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A
-# PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT
-# HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
-# OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE
-# SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
 import logging
 import sys
@@ -17,8 +11,8 @@ from config import AppConfig
 
 conf = AppConfig.load(sys.argv[1])
 
-import apple
-from builder import PageBuilder
+from . import apple
+from .builder import PageBuilder
 
 log = logging.getLogger(__name__)
